requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-graphql-jwt==0.3.1
django-import-export==2.3.0
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
promise==2.3
psycopg2==2.7.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
